Supplicant Enhancements
When `CheckRespawnSupplicant` determines  that a new supplicant is needed, the sprite and starting traits for the supplicant are gathered from the sub god, and cached in the god.
This means that the sprite will not change as the seasons do for a live supplicant, and that ignoring level-up across a season boundary will not result in changed first level trait options.

Moved .gitignore to expected folder.